test: split org writer compat contracts
